Move tests around to be more organized
